requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
confluent-kafka==1.9.2
fastavro==1.7.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-bigquery==3.4.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
numpy==1.24.0
packaging==21.3
pandas==1.5.2
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
